+++ R/print_beast2_options.R
@@ -2,7 +2,7 @@
 #' @inheritParams default_params_doc
 #' @export
 print_beast2_options <- function(beast2_options) {
-  check_beast2_options(beast2_options) # nolint beastier function
+  beastier::check_beast2_options(beast2_options)
   df <- beast2_options_to_table(beast2_options) # nolint beastier function
   cat(knitr::kable(df), sep = "\n")
 }