+++ tests/testthat/test-get_default_beast2_working_dir.R
@@ -1,0 +1,6 @@
+test_that("use", {
+  expect_equal(
+    get_default_beast2_working_dir(),
+    file.path(rappdirs::user_cache_dir(), "beastier")
+  )
+})